Extract BBoxComputationMethod to bboxes module
Move the BBoxComputationMethod enum out of deeplabcut/pose_estimation_pytorch/data/base.py into a new deeplabcut/pose_estimation_pytorch/data/bboxes.py to centralize bbox-related types. Update imports in cocoloader.py, dlcloader.py and base.py to reference the new module, and adjust tests to assert against the enum members instead of string literals. Also remove the now-unused Enum import from base.py.
