+++ tests/pose_estimation_pytorch/data/test_bbox.py
@@ -19,6 +19,7 @@ import pytest
 
 import deeplabcut.pose_estimation_pytorch.data.base as base_mod
 from deeplabcut.pose_estimation_pytorch.data.base import Loader
+from deeplabcut.pose_estimation_pytorch.data.bboxes import BBoxComputationMethod
 from deeplabcut.pose_estimation_pytorch.data.dataset import PoseDatasetParameters
 from deeplabcut.pose_estimation_pytorch.data.dlcloader import DLCLoader
 from deeplabcut.pose_estimation_pytorch.data.utils import bbox_from_keypoints
@@ -161,8 +162,8 @@ def test_dlcloader_default_bbox_method_is_backward_compatible():
     """
     loader = object.__new__(DLCLoader)
 
-    assert DLCLoader.default_bbox_method(loader, Task.TOP_DOWN) == "keypoints"
-    assert DLCLoader.default_bbox_method(loader, Task.DETECT) == "keypoints"
+    assert DLCLoader.default_bbox_method(loader, Task.TOP_DOWN) == BBoxComputationMethod.KEYPOINTS
+    assert DLCLoader.default_bbox_method(loader, Task.DETECT) == BBoxComputationMethod.KEYPOINTS
     assert DLCLoader.default_bbox_method(loader, Task.BOTTOM_UP) is None
